+++ glide-loader/src/main/java/com/ouattararomuald/slider/loaders/glide/GlideImageLoaderFactory.kt
@@ -1,21 +1,17 @@
 package com.ouattararomuald.slider.loaders.glide
 
-import android.widget.ImageView
-import androidx.annotation.DrawableRes
+import com.bumptech.glide.request.RequestOptions
 import com.ouattararomuald.slider.ImageLoader
 
 /**
  * An [ImageLoader.Factory] which uses Glide.
  *
- * @property errorResId an error drawable to be used if the request image could not be loaded.
- * @property placeholderResId a placeholder drawable to be used while the image is being loaded.
- * If the requested image is not immediately available in the memory cache then this resource will
- * be set on the target [ImageView].
+ * @property requestOptions options to customize loads with Glide.
  */
 class GlideImageLoaderFactory(
-  @DrawableRes private val errorResId: Int = 0,
-  @DrawableRes private val placeholderResId: Int = 0
+  private val eventListener: ImageLoader.EventListener? = null,
+  private val requestOptions: RequestOptions? = null
 ) : ImageLoader.Factory<GlideImageLoader> {
 
-  override fun create(): GlideImageLoader = GlideImageLoader(errorResId, placeholderResId)
+  override fun create(): GlideImageLoader = GlideImageLoader(requestOptions, eventListener)
 }