+++ sample/src/main/java/com/slider/demo/ImageLoaderType.kt
@@ -1,6 +1,7 @@
 package com.slider.demo
 
 enum class ImageLoaderType {
+  COIL,
   GLIDE,
   PICASSO
 }