+++ slider-library/src/main/java/com/ouattararomuald/slider/ImageLoader.kt
@@ -3,10 +3,26 @@ package com.ouattararomuald.slider
 import android.widget.ImageView
 
 /** Loads image into [ImageView]. */
-interface ImageLoader {
+abstract class ImageLoader(private val eventListener: EventListener? = null) {
+
+  /** Configures the [imageView] at the given position. */
+  internal fun configureImageView(imageView: ImageView) {
+    eventListener?.onImageViewConfiguration(imageView)
+  }
 
   /** Loads an image into the given [imageView] using the specified path. */
-  fun load(path: String, imageView: ImageView)
+  abstract fun load(path: String, imageView: ImageView)
+
+  interface EventListener {
+    /**
+     * Called before the image is loaded in the [imageView].
+     *
+     * You can use this method to customize the image view.
+     *
+     * @param imageView [ImageView] that will be used to display the image.
+     */
+    fun onImageViewConfiguration(imageView: ImageView)
+  }
 
   /** Creates an [ImageLoader] instance. */
   interface Factory<T : ImageLoader> {